Close button: extensibility in post editor (#22323)
Convert post editor's close button into an extensible slot.
