+++ packages/edit-site/src/index.js
@@ -29,5 +29,4 @@ export function initialize( id, settings ) {
 	render( <Editor settings={ settings } />, document.getElementById( id ) );
 }
 
-export { default as __experimentalMainDashboardButton } from './components/header/main-dashboard-button';
-export { default as __experimentalFullscrenModeClose } from './components/header/fullscreen-mode-close';
+export { default as __experimentalFullscreenModeClose } from './components/header/fullscreen-mode-close';
