+++ packages/interface/src/components/index.js
@@ -2,3 +2,4 @@ export { default as ComplementaryArea } from './complementary-area';
 export { default as FullscreenMode } from './fullscreen-mode';
 export { default as InterfaceSkeleton } from './interface-skeleton';
 export { default as PinnedItems } from './pinned-items';
+export { default as __experimentalMainDashboardButton } from './main-dashboard-button';
